requirements.txt
biopython==1.76
gffutils==0.10.1
pandas==1.0.1
pybedtools==0.8.1
pysam>=0.15
ssw-py
